% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestral_inference.R
\name{ancestor_inference}
\alias{ancestor_inference}
\title{Infer ancestral lineage barcodes and cell states}
\usage{
ancestor_inference(tree, N_char, barcodes, cell_state_labels, state_lineages)
}
\arguments{
\item{tree}{A sub-cell division tree.}

\item{N_char}{Number of characters in the lineage barcode.}

\item{barcodes}{Lineage barcodes of the leaf nodes.}

\item{cell_state_labels}{Cell states of the leaf nodes.}

\item{state_lineages}{A list of state lineages.}
}
\value{
A list containing ancestral lineage barcodes and cell states.
}
\description{
Infer ancestral lineage barcodes and cell states
}
