% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestral_inference.R
\name{ancestor_inference_time_scale}
\alias{ancestor_inference_time_scale}
\title{Infer ancestral barcodes with time scaling}
\usage{
ancestor_inference_time_scale(tree, N_char, barcodes)
}
\arguments{
\item{tree}{A sub-cell division tree.}

\item{N_char}{Total number of sites.}

\item{barcodes}{A matrix of lineage barcodes for the leaf nodes.}
}
\value{
A matrix of barcodes for both leaf and internal nodes, where each internal node's barcode is inferred from its children.
}
\description{
Infer ancestral barcodes with time scaling
}
