% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calculation.R
\name{barcode_similarity_likelihood}
\alias{barcode_similarity_likelihood}
\title{Calculate barcode shift probability}
\usage{
barcode_similarity_likelihood(tree, N_char, edges, barcodes)
}
\arguments{
\item{tree}{Sub-cell division tree.}

\item{N_char}{Total number of sites.}

\item{edges}{Edges of the sub-cell division tree.}

\item{barcodes}{Lineage barcodes.}
}
\value{
Score of barcode shifts.
}
\description{
Calculate barcode shift probability
}
