% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_imputation.R
\name{barcodes_grouping}
\alias{barcodes_grouping}
\title{Group label barcodes (For imputation)}
\usage{
barcodes_grouping(label_barcodes, remain_sites, n)
}
\arguments{
\item{label_barcodes}{A matrix of label barcodes.}

\item{remain_sites}{A matrix of remaining sites.}

\item{n}{If n==1, return grouped label barcodes; otherwise, return grouped remaining sites.}
}
\value{
A list of grouped barcodes.
}
\description{
Group label barcodes (For imputation)
}
