% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{common_mutation_matrix}
\alias{common_mutation_matrix}
\title{Build a common mutation matrix between root barcodes of sub-subtrees}
\usage{
common_mutation_matrix(subtrees_rootbar)
}
\arguments{
\item{subtrees_rootbar}{A list of root barcodes for sub-subtrees.}
}
\value{
A matrix of common mutation counts between each pair of sub-subtree root barcodes.
}
\description{
Build a common mutation matrix between root barcodes of sub-subtrees
}
