% Generated by roxygen2: do not edit by hand

\name{composite_score}
\alias{composite_score}
\title{Calculate composition tree score based on cell state and barcode shift scores}
\usage{
composite_score(
  tree,
  barcodes,
  N_char,
  cell_state_labels,
  state_lineages,
  state_score = NULL,
  barcode_score = NULL,
  lambda_1 = lambda1,
  lambda_2 = lambda2
)
}
\arguments{
\item{tree}{Sub-cell division tree.}

\item{barcodes}{Lineage barcodes.}

\item{N_char}{Total number of sites.}

\item{cell_state_labels}{Vector of cell state labels.}

\item{state_lineages}{List of state lineages.}

\item{state_score}{Precomputed cell state score (optional).}

\item{barcode_score}{Precomputed barcode score (optional).}

\item{lambda_1}{Weight for the cell state score.}

\item{lambda_2}{Weight for the barcode score.}
}
\value{
composition score of cell state and barcode shifts.
}
\description{
Calculate composition tree score based on cell state and barcode shift scores
}
