% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{count_common_mutations}
\alias{count_common_mutations}
\title{Count common mutations between two barcodes}
\usage{
count_common_mutations(barcode1, barcode2, N_char = NULL)
}
\arguments{
\item{barcode1}{A vector representing the first barcode.}

\item{barcode2}{A vector representing the second barcode.}

\item{N_char}{Optional. Number of sites to consider.}
}
\value{
Count of sites with the same mutation (excluding zeros).
}
\description{
Count common mutations between two barcodes
}
