% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{count_uncommon_mutations}
\alias{count_uncommon_mutations}
\title{Count uncommon mutations between two barcodes}
\usage{
count_uncommon_mutations(barcode1, barcode2, N_char = NULL)
}
\arguments{
\item{barcode1}{A vector representing the first barcode.}

\item{barcode2}{A vector representing the second barcode.}

\item{N_char}{Optional. Number of sites to consider.}
}
\value{
Count of sites with different mutations (both nonzero).
}
\description{
Count uncommon mutations between two barcodes
}
