% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{decompose_subtrees}
\alias{decompose_subtrees}
\title{Generate sub-subtrees from best subtrees}
\usage{
decompose_subtrees(bestsubtree, state, barcodes, l_sl)
}
\arguments{
\item{bestsubtree}{A list of best subtrees.}

\item{state}{Data frame with cell state information.}

\item{barcodes}{Matrix of lineage barcodes.}

\item{l_sl}{Length (number) of state lineages.}
}
\value{
A vector (list) of sub-subtrees.
}
\description{
Generate sub-subtrees from best subtrees
}
