% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distance_with_dropout}
\alias{distance_with_dropout}
\title{Define Hamming distance between lineage barcodes with dropout (For imputation)}
\usage{
distance_with_dropout(barcode_1, barcode_2, N_char = NULL)
}
\arguments{
\item{barcode_1}{A character vector representing the first barcode.}

\item{barcode_2}{A character vector representing the second barcode.}

\item{N_char}{Optional. Number of positions to consider in the barcode.}
}
\value{
The Hamming distance between the two barcodes with dropout.
}
\description{
Define Hamming distance between lineage barcodes with dropout (For imputation)
}
