% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{drop_duplicated_tips}
\alias{drop_duplicated_tips}
\title{Drop duplicated leaf nodes from tree list}
\usage{
drop_duplicated_tips(
  tree_list,
  barcodes,
  cell_lineages,
  state_lineage,
  alpha,
  beta
)
}
\arguments{
\item{tree_list}{List of raw sub-cell division trees.}

\item{barcodes}{Matrix of lineage barcodes.}

\item{cell_lineages}{Cell lineages (unused in the current function but kept for compatibility).}

\item{state_lineage}{A list where each element is a vector of state lineage labels for a cell.}

\item{alpha}{Alpha parameter for computing sibling distance.}

\item{beta}{Beta parameter for computing sibling distance.}
}
\value{
Modified tree_list with duplicated leaf nodes dropped.
}
\description{
Drop duplicated leaf nodes from tree list
}
