% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{drop_non_minimum_tips}
\alias{drop_non_minimum_tips}
\title{Drop non-minimum tips from tree list}
\usage{
drop_non_minimum_tips(tree_list, dist_list, cell)
}
\arguments{
\item{tree_list}{List of trees.}

\item{dist_list}{Data frame with columns "lineage" and "dist" containing distances.}

\item{cell}{Cell ID (tip) to be dropped.}
}
\value{
Modified tree_list after dropping the specified tip from trees not in the best (minimum-distance) lineage.
}
\description{
Drop non-minimum tips from tree list
}
