% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_imputation.R
\name{dropout_imputation}
\alias{dropout_imputation}
\title{Impute stochastic dropout of lineage barcodes}
\usage{
dropout_imputation(barcodes_dropout, N_char, r_n)
}
\arguments{
\item{barcodes_dropout}{Lineage barcodes with stochastic dropout.}

\item{N_char}{Total number of sites.}

\item{r_n}{Number of sites used as label barcodes.}
}
\value{
Lineage barcodes after imputation.
}
\description{
Impute stochastic dropout of lineage barcodes
}
