% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_imputation.R
\name{dropout_imputation_alter}
\alias{dropout_imputation_alter}
\title{Impute stochastic dropout in lineage barcodes}
\usage{
dropout_imputation_alter(barcodes_dropout, N_char, ncells, r_n)
}
\arguments{
\item{barcodes_dropout}{A matrix of lineage barcodes with dropout values.}

\item{N_char}{Number of characters (sites) in the barcode.}

\item{ncells}{Number of cells.}

\item{r_n}{Number of selected sites for reference.}
}
\value{
A matrix of lineage barcodes after imputation.
}
\description{
Impute stochastic dropout in lineage barcodes
}
