% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_node_depth}
\alias{find_node_depth}
\title{Find the depth of a given node in a tree}
\usage{
find_node_depth(tree, node, ncells)
}
\arguments{
\item{tree}{A tree object containing "edge" and "edge.length" elements.}

\item{node}{The node (integer) whose depth is to be computed.}

\item{ncells}{Number of leaf nodes in the tree; the root is assumed to be ncells + 1.}
}
\value{
The depth (sum of edge lengths) from the given node to the root.
}
\description{
Find the depth of a given node in a tree
}
