% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_tree_newick}
\alias{format_tree_newick}
\title{Convert tree information into Newick (.nwk) format}
\usage{
format_tree_newick(root_node, edges)
}
\arguments{
\item{root_node}{The current root node.}

\item{edges}{A two-column matrix or data frame of edges (parent in column 1, child in column 2).}
}
\value{
A character string in Newick format.
}
\description{
Convert tree information into Newick (.nwk) format
}
