% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_tree_newick_time}
\alias{format_tree_newick_time}
\title{Convert tree information into Newick (.nwk) format including time information}
\usage{
format_tree_newick_time(root_node, edges, Time)
}
\arguments{
\item{root_node}{The current root node.}

\item{edges}{A two-column matrix or data frame of edges (parent in column 1, child in column 2).}

\item{Time}{A vector of time values corresponding to vertices; vertex names are taken from V(graph)$name.}
}
\value{
A character string in Newick format with branch lengths (time information).
}
\description{
Convert tree information into Newick (.nwk) format including time information
}
