% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{get_first_element}
\alias{get_first_element}
\title{Get first element (actually returns the second element)}
\usage{
get_first_element(x)
}
\arguments{
\item{x}{A list.}
}
\value{
The second element of x.
}
\description{
Get first element (actually returns the second element)
}
