% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{get_root_barcode}
\alias{get_root_barcode}
\title{Get the root barcode from a tree}
\usage{
get_root_barcode(x)
}
\arguments{
\item{x}{A tree object with a root.barcode element.}
}
\value{
The root barcode.
}
\description{
Get the root barcode from a tree
}
