% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{get_subtree_celltypes}
\alias{get_subtree_celltypes}
\title{Get subtree cell types}
\usage{
get_subtree_celltypes(tree_groups)
}
\arguments{
\item{tree_groups}{A list containing two groups of trees. For the first group, each element is a tree object;
for the second group, each element is an index referring to a tree in the global "subsubtrees" list.}
}
\value{
A list of cell type vectors for each subtree, with names "subtree_1", "subtree_2", etc.
}
\description{
Get subtree cell types
}
