% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{get_subtree_root_barcodes}
\alias{get_subtree_root_barcodes}
\title{Get subtree root barcodes for each best subtree in a lineage}
\usage{
get_subtree_root_barcodes(bestsubtree, state, barcodes, l_sl)
}
\arguments{
\item{bestsubtree}{A list of best subtrees.}

\item{state}{Data frame with cell state information.}

\item{barcodes}{Matrix of lineage barcodes.}

\item{l_sl}{Length (number) of state lineages.}
}
\value{
A vector of root barcodes for the sub-subtrees.
}
\description{
Get subtree root barcodes for each best subtree in a lineage
}
