% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{get_tips}
\alias{get_tips}
\title{Get tip labels from a tree}
\usage{
get_tips(x)
}
\arguments{
\item{x}{A tree object.}
}
\value{
A vector of tip labels.
}
\description{
Get tip labels from a tree
}
