% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypic_dynamics_analysis.R
\name{getDescendants_iter}
\alias{getDescendants_iter}
\title{Get all descendants of a node in a phylogenetic tree (iterative)}
\usage{
getDescendants_iter(tree, node)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{"phylo"}.}

\item{node}{An integer giving the node number from which to collect descendants.}
}
\value{
An integer vector of all descendant node/leaf numbers (excluding the input node).
}
\description{
Get all descendants of a node in a phylogenetic tree (iterative)
}
