% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{group_subcell_trees}
\alias{group_subcell_trees}
\title{Group sub-cell division trees for merging based on common mutations}
\usage{
group_subcell_trees(
  CM,
  Nodes_rank,
  Nodes_weight,
  subsubtrees,
  bind_tree_list,
  nsubtree
)
}
\arguments{
\item{CM}{Common mutation matrix.}

\item{Nodes_rank}{Matrix of nodes ranking.}

\item{Nodes_weight}{Matrix of nodes weight.}

\item{subsubtrees}{List of sub-subtrees.}

\item{bind_tree_list}{List of trees to bind.}

\item{nsubtree}{Current number/index for subtrees.}
}
\value{
A list containing the merged tree group and remaining nodes.
}
\description{
Group sub-cell division trees for merging based on common mutations
}
