% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Compute Hamming Distance Matrix for Barcodes}
\usage{
hamming_distance(barcodes)
}
\arguments{
\item{barcodes}{A matrix where each row represents a barcode.}
}
\value{
A Hamming distance matrix computed from the input barcodes.
}
\description{
Compute Hamming Distance Matrix for Barcodes
}
