% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{improved_hamming_distance}
\alias{improved_hamming_distance}
\title{Calculate improved Hamming distance between two lineage barcodes}
\usage{
improved_hamming_distance(barcode1, barcode2, alpha, beta, N_char = NULL)
}
\arguments{
\item{barcode1}{A vector representing the first lineage barcode.}

\item{barcode2}{A vector representing the second lineage barcode.}

\item{alpha}{A positive number; if two sites have the same mutation, the total distance is reduced by alpha.}

\item{beta}{A number greater than 1; if two sites have different mutations, the total distance is increased by beta.}

\item{N_char}{Optional. The number of sites to consider; defaults to the length of barcode1.}
}
\value{
The improved Hamming distance.
}
\description{
Calculate improved Hamming distance between two lineage barcodes
}
