% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_construction.R
\name{initial_tree_construction}
\alias{initial_tree_construction}
\title{Reconstruct initial sub-cell division trees based solely on lineage barcodes}
\usage{
initial_tree_construction(state_lineages, barcodes_lineages)
}
\arguments{
\item{state_lineages}{A list of state lineages.}

\item{barcodes_lineages}{A list of barcode matrices corresponding to each state lineage.}
}
\value{
A list of initial sub-cell division trees reconstructed only from lineage barcodes.
}
\description{
Reconstruct initial sub-cell division trees based solely on lineage barcodes
}
