% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_imputation.R
\name{lable_imputation}
\alias{lable_imputation}
\title{Impute label barcodes}
\usage{
lable_imputation(label)
}
\arguments{
\item{label}{A matrix of label barcodes.}
}
\value{
Label barcodes after imputing stochastic dropout.
}
\description{
Impute label barcodes
}
