% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{merge_subcell_trees_ordered}
\alias{merge_subcell_trees_ordered}
\title{Merge sub-cell division trees in order based on nodes rank and weight}
\usage{
merge_subcell_trees_ordered(
  CM,
  Nodes_rank,
  Nodes_weight,
  subsubtrees,
  bind_tree_list,
  nsubtree
)
}
\arguments{
\item{CM}{Common mutation matrix.}

\item{Nodes_rank}{Matrix of nodes ranking.}

\item{Nodes_weight}{Matrix of nodes weight.}

\item{subsubtrees}{List of sub-subtrees.}

\item{bind_tree_list}{List of trees to bind.}

\item{nsubtree}{Current number/index for subtrees.}
}
\value{
A merged cell division tree.
}
\description{
Merge sub-cell division trees in order based on nodes rank and weight
}
