% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{merge_subcell_trees_ward}
\alias{merge_subcell_trees_ward}
\title{Merge sub-cell division trees using Ward's method}
\usage{
merge_subcell_trees_ward(subtrees_rootbar, subsubtrees)
}
\arguments{
\item{subtrees_rootbar}{A list of root barcodes for sub-subtrees.}

\item{subsubtrees}{A list of sub-subtrees.}
}
\value{
A merged cell division tree.
}
\description{
Merge sub-cell division trees using Ward's method
}
