% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_construction.R
\name{nodes_clustering}
\alias{nodes_clustering}
\title{Generate a tree topology using Ward's clustering method}
\usage{
nodes_clustering(distance_matrix)
}
\arguments{
\item{distance_matrix}{A Hamming distance matrix.}
}
\value{
A phylogenetic tree obtained by clustering using Ward's method, with all edge lengths set to 1.
}
\description{
Generate a tree topology using Ward's clustering method
}
