% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_refinement.R
\name{nodes_sampling}
\alias{nodes_sampling}
\title{Calculate the probability of each node being selected as the root for subtree exchange}
\usage{
nodes_sampling(tree, edges, cell_state_labels, barcodes, state_lineages)
}
\arguments{
\item{tree}{Sub-cell division tree.}

\item{edges}{Edge matrix of the sub-cell division tree.}

\item{cell_state_labels}{Vector of cell state labels.}

\item{barcodes}{Matrix of lineage barcodes.}

\item{state_lineages}{List of state lineage paths.}
}
\value{
A probability vector for each node.
}
\description{
Calculate the probability of each node being selected as the root for subtree exchange
}
