% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{phylogenetic_tree}
\alias{phylogenetic_tree}
\title{Compute a time-scaled phylogenetic tree based on lineage barcodes}
\usage{
phylogenetic_tree(
  tree,
  N_char,
  barcodes,
  ncells = NULL,
  Nnodes = NULL,
  edges = NULL
)
}
\arguments{
\item{tree}{A sub-cell division tree.}

\item{N_char}{Total number of sites.}

\item{barcodes}{A matrix of lineage barcodes.}

\item{ncells}{Optional. Number of cells (defaults to the number of tip labels in the tree).}

\item{Nnodes}{Optional. Number of nodes (defaults to 2 * ncells - 2).}

\item{edges}{Optional. Edge information (ignored if NULL).}
}
\value{
A time-scaled phylogenetic tree.
}
\description{
Compute a time-scaled phylogenetic tree based on lineage barcodes
}
