% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{prune_cell_tree}
\alias{prune_cell_tree}
\title{Subdivide a sub-cell division tree into sub-subtrees}
\usage{
prune_cell_tree(sub_tree, state, barcodes, subid, itr_time)
}
\arguments{
\item{sub_tree}{A sub-cell division tree.}

\item{state}{A data frame with cell state information.}

\item{barcodes}{A matrix of lineage barcodes.}

\item{subid}{Identifier for the sub-cell division tree.}

\item{itr_time}{Iteration time (for recursive calls).}
}
\value{
A list of sub-subtrees.
}
\description{
Subdivide a sub-cell division tree into sub-subtrees
}
