% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_imputation.R
\name{remaining_sites_imputation}
\alias{remaining_sites_imputation}
\title{Impute remaining sites}
\usage{
remaining_sites_imputation(remain_s, alpha)
}
\arguments{
\item{remain_s}{A matrix of remaining sites.}

\item{alpha}{Threshold for dropout ratio to decide if imputation should occur.}
}
\value{
Remaining sites after imputing stochastic dropout.
}
\description{
Impute remaining sites
}
