% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{sibling_cells_distance}
\alias{sibling_cells_distance}
\title{Find the distance between sibling cells}
\usage{
sibling_cells_distance(cell, tree, lineage, alpha, beta)
}
\arguments{
\item{cell}{A tip label (cell ID) for which to find the sibling distance.}

\item{tree}{A sub-cell division tree.}

\item{lineage}{The state lineage identifier (e.g., "L1") that the cell belongs to.}

\item{alpha}{Alpha parameter for the improved Hamming distance.}

\item{beta}{Beta parameter for the improved Hamming distance.}
}
\value{
The improved Hamming distance between the sibling cells.
}
\description{
Find the distance between sibling cells
}
