% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_tree_inference.R
\name{state_lineage_info}
\alias{state_lineage_info}
\title{State lineage information}
\usage{
state_lineage_info(state_lineages, ncells, state, barcodes)
}
\arguments{
\item{state_lineages}{A named list of state lineage paths (e.g., "L1", "L2", ...).}

\item{ncells}{Number of cells.}

\item{state}{A data frame containing cell states (with columns "cluster" and "cell_id").}

\item{barcodes}{A matrix of lineage barcodes for each cell.}
}
\value{
A list with the following elements:
\describe{
\item{state_lineage}{A list indicating which state lineage(s) each cell belongs to.}
\item{cell_lineages}{A list mapping each state lineage to the corresponding cell IDs.}
\item{barcodes_lineages}{A list of barcode matrices for cells in each state lineage.}
\item{state_labels_lineages}{A list of state labels (clusters) for cells in each state lineage.}
}
}
\description{
State lineage information
}
