% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calculation.R
\name{state_transition_likelihood}
\alias{state_transition_likelihood}
\title{Calculate state shift probability for cell states}
\usage{
state_transition_likelihood(
  tree,
  edges,
  barcodes,
  cell_state_labels,
  state_lineages
)
}
\arguments{
\item{tree}{Sub-cell division tree.}

\item{edges}{Edges of the sub-cell division tree.}

\item{barcodes}{Lineage barcodes.}

\item{cell_state_labels}{Vector of cell state labels.}

\item{state_lineages}{List of state lineages.}
}
\value{
Score of cell state shifts.
}
\description{
Calculate state shift probability for cell states
}
