% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{stochastic_dropout}
\alias{stochastic_dropout}
\title{Stochastic dropout generation}
\usage{
stochastic_dropout(barcodes_initial, p_ds)
}
\arguments{
\item{barcodes_initial}{A matrix of barcodes containing only inheritable dropout.}

\item{p_ds}{Stochastic dropout rate.}
}
\value{
The initial barcodes with stochastic dropouts applied.
}
\description{
Generate stochastic dropouts on lineage barcodes.
}
