% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_refinement.R
\name{subtree_refinement}
\alias{subtree_refinement}
\title{Refine sub-cell division trees based on lineage barcodes and cell states}
\usage{
subtree_refinement(
  Trees_initial,
  state_lineages,
  barcodes_lineages,
  N_char,
  state_labels_lineages,
  lambda1,
  lambda2,
  maxIter = 100,
  repeat_time = 10
)
}
\arguments{
\item{Trees_initial}{List of initial sub-cell division trees (by state lineage).}

\item{state_lineages}{List of state lineage paths.}

\item{barcodes_lineages}{List of barcode matrices per state lineage.}

\item{N_char}{Total number of sites.}

\item{state_labels_lineages}{List of state labels per state lineage.}

\item{lambda1}{Weight for cell state score.}

\item{lambda2}{Weight for barcode score.}

\item{maxIter}{Maximum number of iterations.}

\item{repeat_time}{Number of iterations required for convergence.}
}
\value{
A list containing:
\describe{
\item{bestsubtree}{List of refined sub-cell division trees (one per state lineage).}
\item{bestsubtreescore}{List of corresponding best scores for each refined tree.}
\item{total_time}{Total running time.}
}
}
\description{
Refine sub-cell division trees based on lineage barcodes and cell states
}
