% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{subtree_root_bar}
\alias{subtree_root_bar}
\title{Get root barcodes from a list of trees}
\usage{
subtree_root_bar(x)
}
\arguments{
\item{x}{A list of tree objects.}
}
\value{
A list of root barcodes.
}
\description{
Get root barcodes from a list of trees
}
