% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_integration.R
\name{subtrees_rank}
\alias{subtrees_rank}
\title{Rank subtrees based on common mutation counts}
\usage{
subtrees_rank(CM)
}
\arguments{
\item{CM}{A common mutation matrix.}
}
\value{
A list containing the nodes rank matrix and nodes weight matrix.
}
\description{
Rank subtrees based on common mutation counts
}
