% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_refinement.R
\name{subtrees_swapping}
\alias{subtrees_swapping}
\title{Swap subtrees in a sub-cell division tree}
\usage{
subtrees_swapping(
  tree,
  edges,
  cell_state_labels,
  barcodes,
  state_lineages,
  tree_renew_prob = NULL
)
}
\arguments{
\item{tree}{Sub-cell division tree.}

\item{edges}{Edge matrix of the tree.}

\item{cell_state_labels}{Vector of cell state labels.}

\item{barcodes}{Matrix of lineage barcodes.}

\item{state_lineages}{List of state lineage paths.}

\item{tree_renew_prob}{Optional probability vector for selecting nodes for subtree exchange.}
}
\value{
The tree after swapping selected subtrees.
}
\description{
Swap subtrees in a sub-cell division tree
}
